run the project from the build/install directory
We want to run the program from inside the install directory because that is the place where it would run on the user's system.
This has the advantage that log files are no longer in the project directory but rather in the build directory and will be deleted with the clean task.