FIx a small concurrency problem here.
It is possible that the extension will clear the table model entries in between the draw thread calculating the loop limits and calling the cell render.
This can cause the TableModel to throw a null pointer exception when it tries to get an entry for a row that doesnt exist anymore.