Avoid a crash when clicking on the first pixel column.
A Component or Container would return null to a click when that click
was at the first column for that component and would cause a nullpointer
exception down the line.

Make the click calculation inclusive to the first column to avoid this.
Also reorder the LPBase method to make sure that a call to
Component::setFocused is always non null.