Avoid NPE when track data event is delayed
Currently the track data event is not published for the Nordschleife.
This fix means that the gap calculation does not work however,
it is impossible to work without the track data. This is an issue
that needs to be adressed by kunos.
This is a simple fix to avoid a npe to allow the program to keep
running. Other features that require the track data already are
resiliant again this. The gap calculator really should be more
resiliant as well but this is good enough for now.